refactor: loader 컴포넌트 이동
